One line code added, one row in data set removed
Needed to make change to code. I noticed that Excel uses first row to hold column names, but R does not. This causes R to have an additional row at the bottom of the data set with NA values. I included a line of code after bringing in the data set to mitigate this issue.